Ch51Ex29 - незавершенное
