+++ src/main/java/ru/kulichenkom/ekkel/chapter15/Holder.java
@@ -1,0 +1,24 @@
+package ru.kulichenkom.ekkel.chapter15;
+
+public class Holder<T> {
+    private T value;
+
+    public Holder() {
+    }
+
+    public Holder(T val) {
+        value = val;
+    }
+
+    public void set(T val) {
+        value = val;
+    }
+
+    public T get() {
+        return value;
+    }
+
+    public boolean equals(Object obj) {
+        return value.equals(obj);
+    }
+}
