+++ src/main/java/ru/kulichenkom/ekkel/chapter15/UnboundedWildcards.java
@@ -91,28 +91,6 @@ class UnboundedWildcards2 {
 
 //: generics/Holder.java
 
-class Holder<T> {
-    private T value;
-
-    public Holder() {
-    }
-
-    public Holder(T val) {
-        value = val;
-    }
-
-    public void set(T val) {
-        value = val;
-    }
-
-    public T get() {
-        return value;
-    }
-
-    public boolean equals(Object obj) {
-        return value.equals(obj);
-    }
-}
 //: generics/Wildcards.java
 // Exploring the meaning of wildcards.
